Split into ereader_helpers.c, faraway_island.c, and ereader_screen.c
