+++ data/ereader_screen.s
@@ -1,0 +1,7 @@
+	.include "asm/macros.inc"
+	.include "constants/constants.inc"
+
+	.section .rodata
+
+gUnknown_0862609C:: @ 862609C
+	.incbin "data/unknown_jp_62609C.bin"
