+++ include/faraway_island.h
@@ -1,11 +1,10 @@
-#ifndef GUARD_EREADER_H
-#define GUARD_EREADER_H
+#ifndef GUARD_FARAWAY_ISLAND_H
+#define GUARD_FARAWAY_ISLAND_H
 
-bool32 sub_81D3B34(void);
 u32 sub_81D427C(void);
 bool8 sub_81D4A58(struct EventObject*);
 void UpdateFarawayIslandStepCounter(void);
 bool8 EventObjectIsFarawayIslandMew(struct EventObject *);
 bool8 IsMewPlayingHideAndSeek(void);
 
-#endif // GUARD_EREADER_H
+#endif // GUARD_FARAWAY_ISLAND_H
