+++ src/event_object_movement.c
@@ -3,10 +3,10 @@
 #include "battle_pyramid.h"
 #include "berry.h"
 #include "decoration.h"
-#include "ereader.h"
 #include "event_data.h"
 #include "event_object_movement.h"
 #include "event_scripts.h"
+#include "faraway_island.h"
 #include "field_camera.h"
 #include "field_effect.h"
 #include "field_effect_helpers.h"
