+++ src/trainer_hill.c
@@ -3,7 +3,7 @@
 #include "battle.h"
 #include "battle_tower.h"
 #include "battle_setup.h"
-#include "ereader.h"
+#include "ereader_helpers.h"
 #include "event_data.h"
 #include "event_scripts.h"
 #include "fieldmap.h"
